requirements.txt
appdirs==1.4.4
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
behave==1.2.6
blinker==1.7.0
Brotli==1.1.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cryptography==42.0.2
cssselect==1.2.0
cycler==0.12.1
Cython==0.29.37
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
execnet==2.0.2
executing==2.0.1
fake-useragent==1.4.0
fasteners==0.19
filelock==3.13.1
Flask==3.0.2
fonttools==4.49.0
frozendict==2.4.0
fsspec==2024.2.0
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hdbscan==0.8.33
hpack==4.0.0
html5lib==1.1
hyperframe==6.0.1
idna==3.6
iniconfig==2.0.0
ipykernel==6.29.2
ipython==8.21.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kaitaistruct==0.10
kiwisolver==1.4.5
llvmlite==0.42.0
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multitasking==0.0.11
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
parameterized==0.9.0
parse==1.20.1
parse-type==0.6.2
parso==0.8.3
patsy==0.5.6
pdbp==1.5.0
peewee==3.17.1
pexpect==4.9.0
phonenumbers==8.13.30
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==15.0.0
pyasn1==0.5.1
pycparser==2.21
Pygments==2.17.2
pynndescent==0.5.11
pynose==1.4.8
pyOpenSSL==24.0.0
pyotp==2.9.0
pyparsing==3.1.1
PySocks==1.7.1
pytest==8.0.1
pytest-html==2.0.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==13.0
pytest-xdist==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pythontextnow==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
random-user-agent==1.0.1
requests==2.31.0
rich==13.7.0
sbvirtualdisplay==1.3.0
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
selenium==4.17.2
selenium-wire==5.1.0
seleniumbase==4.23.5
shap==0.44.1
six==1.16.0
sklearn==0.0
slicer==0.0.7
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.1
sympy==1.12
tabcompleter==1.3.0
threadpoolctl==3.3.0
tomli==2.0.1
torch==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.1
umap-learn==0.5.5
undetected-chromedriver==3.5.5
urllib3==2.2.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
xgboost==2.0.3
yfinance==0.2.36
zstandard==0.22.0
